fix: changed the directory of Logbook, Diving, Moim
