--- tangsquad/src/main/java/backend/tangsquad/dto/request/MoimLeaderUpdateRequest.java
@@ -1,13 +1,0 @@
-package backend.tangsquad.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class MoimLeaderUpdateRequest {
-    private Long id;
-    private Long userId;
-}
