--- tangsquad/src/main/java/backend/tangsquad/dto/response/LogListReadResponse.java
@@ -1,17 +1,0 @@
-package backend.tangsquad.dto.response;
-
-import backend.tangsquad.domain.User;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-public class LogListReadResponse {
-    private User user;
-    private List<LogReadResponse> Logs;
-    private int currentPage;
-    private int totalItems;
-    private int totalPages;
-}
