--- tangsquad/src/main/java/backend/tangsquad/dto/response/MoimLeaderUpdateResponse.java
@@ -1,13 +1,0 @@
-package backend.tangsquad.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class MoimLeaderUpdateResponse {
-    private Long id;
-    private Long userId;
-}
