+++ tangsquad/src/main/java/backend/tangsquad/moim/dto/request/MoimLeaderUpdateRequest.java
@@ -1,0 +1,13 @@
+package backend.tangsquad.moim.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class MoimLeaderUpdateRequest {
+    private Long id;
+    private Long userId;
+}
