+++ tangsquad/src/main/java/backend/tangsquad/moim/dto/response/MoimLeaderUpdateResponse.java
@@ -1,0 +1,13 @@
+package backend.tangsquad.moim.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class MoimLeaderUpdateResponse {
+    private Long id;
+    private Long userId;
+}
