+++ tangsquad/src/main/resources/application-API-KEY.properties
@@ -1,0 +1,19 @@
+JWT_SECRET=43181d926de13821cb10c14d8c60c721b3bc3fe86b7afdf6285f56ea994199ed
+
+SPRING_DATASOURCE_PASSWORD=tangsquad123!!
+
+SPRING_DATASOURCE_URL=jdbc:mysql://tangsquad-db.c1ucsomecdfe.ap-northeast-2.rds.amazonaws.com:3306/tangsquad
+
+SPRING_DATASOURCE_USERNAME=admin
+
+SERVER_URL=13.125.41.75
+
+REDIS_PORT=6379
+
+EMAIL_PASSWORD=fvlc laqd nebt acar
+
+COOLSMS_API_KEY=NCSN1NDW3XBBIYGS
+
+COOLSMS_API_SECRET=FYZRBWAXLJCJCP2SPNEBSIX4OABVBSPH
+
+PHONE=01053861475
