+++ tangsquad/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 spring.application.name=tangsquad
-
+spring.config.import=classpath:application-API-KEY.properties
 # Redis Settings
 spring.data.redis.host=${SERVER_URL}
 spring.data.redis.port=${REDIS_PORT}
@@ -43,3 +43,5 @@ spring.mail.properties.mail.smtp.starttls.enable=true
 coolsms.api.key=${COOLSMS_API_KEY}
 coolsms.api.secret=${COOLSMS_API_SECRET}
 coolsms.phone.number=${PHONE}
+
+
